Update pgscen-rts-pca with the tuning option

+++ pgscen/rts_gmlc/pca_command_line.py
@@ -21,10 +21,15 @@ from .data_utils import load_load_data, load_solar_data
 from ..utils.data_utils import (split_actuals_hist_future,
                                 split_forecasts_hist_future)
 
-
 rts_pca_parser = argparse.ArgumentParser(add_help=False, parents=[pca_parser])
 rts_pca_parser.add_argument('rts_dir', type=str,
                             help="where RTS-GMLC repository is stored")
+rts_pca_parser.add_argument('--tuning', '-t', action='store_true',
+                            dest='tuning', help="implement tuning to output scores")
+rts_pca_parser.add_argument('--asset-rho-list', action="extend", nargs="+", type=int,
+                            dest='asset_rho_list', help='the list of asset rho values for tuning')
+rts_pca_parser.add_argument('--time-rho-list', action="extend", nargs="+", type=int,
+                            dest='time_rho_list', help='the list of time rho values for tuning')
 
 joint_parser = argparse.ArgumentParser(parents=[rts_pca_parser],
                                        add_help=False)
@@ -71,23 +76,26 @@ def run_rts_pca() -> None:
     parser = argparse.ArgumentParser(
         'pgscen-rts-pca', parents=[joint_parser],
         description="Create day-ahead RTS load, wind, and solar PCA scenarios."
-        )
+    )
 
     parser.add_argument('--joint', action='store_true',
                         help="use a joint load-solar model")
     args = parser.parse_args()
 
-    rts_runner(args.start, args.days, args.rts_dir, args.out_dir,
-               args.scenario_count, args.nearest_days, args.asset_rho,
-               args.time_rho, args.random_seed, create_load=not args.joint,
-               create_wind=True, write_csv=not args.pickle,
-               skip_existing=args.skip_existing,
-               get_energy_scores=args.energy_scores,
-               get_variograms=args.variograms, verbosity=args.verbose)
+    if args.tuning:
+        for asset_rho in args.asset_rho_list:
+            for time_rho in args.time_rho_list:
+                rts_runner(args.start, args.days, args.rts_dir, args.out_dir,
+                           args.scenario_count, args.nearest_days, asset_rho,
+                           time_rho, args.random_seed, create_load=not args.joint,
+                           create_wind=True, write_csv=not args.pickle,
+                           skip_existing=args.skip_existing,
+                           get_energy_scores=args.energy_scores,
+                           get_variograms=args.variograms, tuning=args.tuning, verbosity=args.verbose)
 
     # to avoid output files from the non-PCA scenarios from being overwritten,
     # we move them to a temporary directory before creating PCA scenarios
-    if args.pickle:
+    if args.pickle & args.tuning:
         os.makedirs(str(Path(args.out_dir, 'tmp')), exist_ok=True)
 
         scen_lbls = ["scens_{}.p.gz".format(start_time.strftime('%Y-%m-%d'))
@@ -123,18 +131,21 @@ def run_rts_pca() -> None:
             os.rename(str(Path(args.out_dir, vrgm_lbl)),
                       str(Path(args.out_dir, 'tmp', vrgm_lbl)))
 
-    rts_pca_runner(args.start, args.days, args.rts_dir, args.out_dir,
-                   args.scenario_count, args.components, args.nearest_days,
-                   args.asset_rho, args.time_rho, args.random_seed,
-                   create_load_solar=args.joint, write_csv=not args.pickle,
-                   skip_existing=args.skip_existing,
-                   get_energy_scores=args.energy_scores,
-                   get_variograms=args.variograms,
-                   use_all_load_hist=args.use_all_load_hist,
-                   verbosity=args.verbose)
+    if args.tuning:
+        for asset_rho in args.asset_rho_list:
+            for time_rho in args.time_rho_list:
+                rts_pca_runner(args.start, args.days, args.rts_dir, args.out_dir,
+                               args.scenario_count, args.components, args.nearest_days,
+                               asset_rho, time_rho, args.random_seed,
+                               create_load_solar=args.joint, write_csv=not args.pickle,
+                               skip_existing=args.skip_existing,
+                               get_energy_scores=args.energy_scores,
+                               get_variograms=args.variograms,
+                               use_all_load_hist=args.use_all_load_hist,
+                               verbosity=args.verbose)
 
     # to create the final output files we merge the PCA and the non-PCA outputs
-    if args.pickle:
+    if args.pickle & args.tuning:
         for scen_lbl in scen_lbls:
             with bz2.BZ2File(Path(args.out_dir, scen_lbl), 'r') as f:
                 pca_scens = pickle.load(f)
@@ -168,6 +179,7 @@ def run_rts_pca() -> None:
                 pickle.dump(pca_scens, f, protocol=-1)
 
 
+
 #TODO: lot of overlap here with the t7k PCA runner
 def rts_pca_runner(start_date: str, ndays: int, rts_dir: str, out_dir: str,
                    scen_count: int, components: str,
@@ -177,7 +189,7 @@ def rts_pca_runner(start_date: str, ndays: int, rts_dir: str, out_dir: str,
                    skip_existing: bool = False,
                    get_energy_scores: bool = False,
                    get_variograms: bool = False,
-                   use_all_load_hist: bool = False,
+                   use_all_load_hist: bool = False, tuning: bool = False,
                    verbosity: int = 0) -> None:
     start = ' '.join([start_date, "08:00:00"])
 
@@ -193,7 +205,7 @@ def rts_pca_runner(start_date: str, ndays: int, rts_dir: str, out_dir: str,
         solar_meta_df) = load_solar_data(rts_dir)
 
     if create_load_solar:
-        load_zone_actual_df, load_zone_forecast_df = load_load_data(rts_dir)
+        load_zone_actual_df, load_zone_forecast_df = loadata(rts_dir)
 
     if verbosity >= 2:
         t0 = time.time()
@@ -201,7 +213,7 @@ def rts_pca_runner(start_date: str, ndays: int, rts_dir: str, out_dir: str,
     # create scenarios for each requested day
     for scenario_start_time in pd.date_range(start=start, periods=ndays,
                                              freq='D', tz='utc'):
-        date_lbl = scenario_start_time.strftime('%Y-%m-%d')
+        date_lbl = scenario_start_time.strft_load_dime('%Y-%m-%d')
 
         scen_timesteps = pd.date_range(start=scenario_start_time,
                                        periods=24, freq='H')
@@ -216,22 +228,23 @@ def rts_pca_runner(start_date: str, ndays: int, rts_dir: str, out_dir: str,
 
         # don't generate scenarios for this day if they have already been saved
         # in this output directory
-        if not write_csv:
-            out_fl = Path(out_dir, "scens_{}.p.gz".format(date_lbl))
+        if not tuning:
+            if not write_csv:
+                out_fl = Path(out_dir, "scens.p.gz".format(date_lbl))
 
-            if skip_existing and out_fl.exists():
-                continue
+                if skip_existing and out_fl.exists():
+                    continue
 
-            out_scens = dict()
+                out_scens = dict()
 
-        if write_csv and skip_existing:
-            date_path = Path(out_dir, scenario_start_time.strftime('%Y%m%d'))
+            if write_csv and skip_existing:
+                date_path = Path(out_dir, scenario_start_time.strftime('%Y%m%d'))
 
-            if create_load_solar and Path(date_path, 'load').exists():
-                shutil.rmtree(Path(date_path, 'load'))
+                if create_load_solar and Path(date_path, 'load').exists():
+                    shutil.rmtree(Path(date_path, 'load'))
 
-            if Path(date_path, 'solar').exists():
-                shutil.rmtree(Path(date_path, 'solar'))
+                if Path(date_path, 'solar').exists():
+                    shutil.rmtree(Path(date_path, 'solar'))
 
         # split input datasets into training and testing subsets
         (solar_site_actual_hists,
@@ -284,21 +297,21 @@ def rts_pca_runner(start_date: str, ndays: int, rts_dir: str, out_dir: str,
                 variograms['Load'] = compute_variograms(
                     solar_engn.scenarios['load'],
                     load_zone_actual_df, load_zone_forecast_df
-                    )
+                )
                 variograms['Solar'] = compute_variograms(
                     solar_engn.scenarios['solar'],
                     solar_site_actual_df, solar_site_forecast_df
-                    )
+                )
+            if not tuning:
+                if write_csv:
+                    solar_engn.write_to_csv(out_dir,
+                                            {'load': load_zone_actual_futures,
+                                             'solar': solar_site_actual_futures},
+                                            write_forecasts=True)
 
-            if write_csv:
-                solar_engn.write_to_csv(out_dir,
-                                        {'load': load_zone_actual_futures,
-                                         'solar': solar_site_actual_futures},
-                                        write_forecasts=True)
-
-            else:
-                out_scens['Load'] = solar_engn.scenarios['load'].round(4)
-                out_scens['Solar'] = solar_engn.scenarios['solar'].round(4)
+                else:
+                    out_scens['Load'] = solar_engn.scenarios['load'].round(4)
+                    out_scens['Solar'] = solar_engn.scenarios['solar'].round(4)
 
             if verbosity >= 2:
                 mdl_components = solar_engn.solar_md.num_of_components
@@ -314,39 +327,39 @@ def rts_pca_runner(start_date: str, ndays: int, rts_dir: str, out_dir: str,
                 energy_scores['Solar'] = compute_energy_scores(
                     solar_engn.scenarios['solar'],
                     solar_site_actual_df, solar_site_forecast_df
-                    )
+                )
 
             if get_variograms:
                 variograms['Solar'] = compute_variograms(
                     solar_engn.scenarios['solar'],
                     solar_site_actual_df, solar_site_forecast_df
-                    )
+                )
+            if not tuning:
+                if write_csv:
+                    solar_engn.write_to_csv(out_dir,
+                                            {'solar': solar_site_actual_futures},
+                                            write_forecasts=True)
 
-            if write_csv:
-                solar_engn.write_to_csv(out_dir,
-                                        {'solar': solar_site_actual_futures},
-                                        write_forecasts=True)
-
-            else:
-                out_scens['Solar'] = solar_engn.scenarios['solar'].round(4)
+                else:
+                    out_scens['Solar'] = solar_engn.scenarios['solar'].round(4)
 
             if verbosity >= 2:
                 mdl_components = solar_engn.model.num_of_components
                 mdl_explained = 1 - solar_engn.model.pca_residual
 
-        if not write_csv:
+        if not write_csv and not tuning:
             with bz2.BZ2File(out_fl, 'w') as f:
                 pickle.dump(out_scens, f, protocol=-1)
 
         if get_energy_scores:
             with bz2.BZ2File(Path(out_fl.parent,
-                                  'escores_{}.p.gz'.format(date_lbl)),
+                                  'escores_{}_{}_{}.p.gz'.format(date_lbl, asset_rho, time_rho)),
                              'w') as f:
                 pickle.dump(energy_scores, f, protocol=-1)
 
         if get_variograms:
             with bz2.BZ2File(Path(out_fl.parent,
-                                  'varios_{}.p.gz'.format(date_lbl)),
+                                  'varios_{}_{}_{}.p.gz'.format(date_lbl, asset_rho, time_rho)),
                              'w') as f:
                 pickle.dump(variograms, f, protocol=-1)
